fix bug of type not match

+++ mindspore/lite/src/runtime/kernel/cpu/base/softmax_base.cc
@@ -44,9 +44,9 @@ int SoftmaxBaseCPUKernel::ReSize() {
   auto in_shape = input_tensor->shape();
   auto in_dims = in_shape.size();
   int ele_size = 1;
-  softmax_param_->n_dim_ = in_dims;
+  softmax_param_->n_dim_ = static_cast<int>(in_dims);
   if (softmax_param_->axis_ == -1) {
-    softmax_param_->axis_ += in_dims;
+    softmax_param_->axis_ += static_cast<int>(in_dims);
   }
   for (size_t i = 0; i < in_dims; i++) {
     softmax_param_->input_shape_[i] = in_shape.at(i);
